Add Subtitles to Video

video.js contains the ability to include tracks with the video, including subtitles, captions and descriptions. includeSubtitles will make sure that they are included on load, and find the defaults to embed with the video.

includeSubtitles(
  video,
  files,
  language = "en",
  label = "English",
  kind = "subtitles",
  default = FALSE
)

Arguments

video

A video()

files

A vector of WebVTT files that contain "cues" of when text should appear, hide and what text to display

language

The valid BCP 47 code for the language of the text track, e.g. "en" for English or "es" for Spanish.

label

Short descriptive text for the track that will used in the user interface. For example, in a menu for selecting a captions language.

kind

An optional vector to match the type of subtitles in files:

subtitles

(default): Translations of the dialogue in the video for when audio is available but not understood. Subtitles are shown over the video.

captions

Transcription of the dialogue, sound effects, musical cues, and other audio information for viewer who are deaf/hard of hearing, or the video is muted. Captions are also shown over the video.

chapters

Chapter titles that are used to create navigation within the video. Typically, these are in the form of a list of chapters that the viewer can use to navigate the video.

descriptions

Text descriptions of the action in the content for when the video portion isn't available or because the viewer is blind or not using a screen. Descriptions are read by a screen reader or turned into a separate audio track.

metadata

Tracks that have data meant for JavaScript to parse and do something with. These aren't shown to the user.

default

The boolean default attribute can be used to indicate that a track's mode should start as "showing". Otherwise, the viewer would need to select their language from a captions or subtitles menu.

Value

An updated video with subtitles included

Details

All vectors must either be the same length as files or of length 1. In the latter, they will be applied to all files supplied.

Examples

vid <- video("https://vjs.zencdn.net/v/oceans.mp4")
includeSubtitles(vid, "url/to/subtitles")

    To view this video please enable JavaScript, and consider upgrading to a web browser that
    supports HTML5 video